refactor(email-2fa): move IMAP client factory to adapter module
